+++ src/camera_service.hpp
@@ -5,6 +5,6 @@ class CameraService {
     uint32_t recording_index{0};
     void startMotionDetection();
 public:
-    static void streamRecording(uint32_t command, uint32_t id);
+    static void streamRecording(uint8_t command, uint8_t id);
     CameraService();
 };
